fix(composer): break annotation component dependency cycle
